requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.14.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
boto3==1.26.89
botocore==1.29.89
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmake==3.25.2
dill==0.3.6
et-xmlfile==1.1.0
exceptiongroup==1.1.0
ffmpeg-python==0.2.0
filelock==3.9.1
flake8==6.0.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Ext==0.1
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
future==0.18.3
futures==3.0.5
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-dialogflow==2.20.0
google-cloud-speech==2.18.0
googleapis-common-protos==1.58.0
googletrans==3.1.0a0
goslate==1.5.4
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.9.0
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.2
mccabe==0.7.0
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
net-tools==0.1.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.27.2
openai-whisper==20230314
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230321
pathspec==0.11.0
pip==23.0.1
platformdirs==3.0.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pydub==0.25.1
pyflakes==3.0.1
pylint==2.16.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2023.2
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soundfile==0.12.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.5.post1
tiktoken==0.3.1
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1
tqdm==4.65.0
triton==2.0.0
types-pytz==2023.2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.38.4
whisper==1.1.10
wrapt==1.14.1
yarl==1.8.2
zipp==3.15.0
